moves LocaleName into cards/components/ folder
