+++ packages/cms/src/components/cards/components/LocaleName.css
@@ -1,4 +1,4 @@
-@import "../../css/mixins.css";
+@import "../../../css/mixins.css";
 
 /* override heading style */
 .cms .cms-locale-name {
